Created InterruptibleMapIterator for system use
* Pulled system code out of SortedMapIterator to create
InterruptibleMapIterator for internal use, allowing SortedMapIterator to
be treated as public API
