+++ core/src/main/java/org/apache/accumulo/core/iterators/SortedMapIterator.java
@@ -22,55 +22,40 @@ import java.util.Iterator;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.SortedMap;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 import org.apache.accumulo.core.client.SampleNotPresentException;
 import org.apache.accumulo.core.data.ByteSequence;
 import org.apache.accumulo.core.data.Key;
 import org.apache.accumulo.core.data.Range;
 import org.apache.accumulo.core.data.Value;
-import org.apache.accumulo.core.iteratorsImpl.system.ColumnFamilySkippingIterator;
-import org.apache.accumulo.core.iteratorsImpl.system.InterruptibleIterator;
 
 /**
  * A simple iterator over a Java SortedMap
  *
  * Note that this class is intended as an in-memory replacement for RFile$Reader, so its behavior
  * reflects the same assumptions; namely, that this iterator is not responsible for respecting the
- * columnFamilies passed into seek(). If you want a Map-backed Iterator that returns only sought
- * CFs, construct a new ColumnFamilySkippingIterator(new SortedMapIterator(map)).
- *
- * @see ColumnFamilySkippingIterator
+ * columnFamilies passed into seek().
  */
-public class SortedMapIterator implements InterruptibleIterator {
+public class SortedMapIterator implements SortedKeyValueIterator<Key,Value> {
   private Iterator<Entry<Key,Value>> iter;
   private Entry<Key,Value> entry;
 
   private SortedMap<Key,Value> map;
   private Range range;
 
-  private AtomicBoolean interruptFlag;
-  private int interruptCheckCount = 0;
-
   @Override
   public SortedMapIterator deepCopy(IteratorEnvironment env) {
     if (env != null && env.isSamplingEnabled()) {
       throw new SampleNotPresentException();
     }
-    return new SortedMapIterator(map, interruptFlag);
+    return new SortedMapIterator(map);
   }
 
-  private SortedMapIterator(SortedMap<Key,Value> map, AtomicBoolean interruptFlag) {
+  public SortedMapIterator(SortedMap<Key,Value> map) {
     this.map = map;
     iter = null;
     this.range = new Range();
     entry = null;
-
-    this.interruptFlag = interruptFlag;
-  }
-
-  public SortedMapIterator(SortedMap<Key,Value> map) {
-    this(map, null);
   }
 
   @Override
@@ -94,9 +79,6 @@ public class SortedMapIterator implements InterruptibleIterator {
     if (entry == null)
       throw new IllegalStateException();
 
-    if (interruptFlag != null && interruptCheckCount++ % 100 == 0 && interruptFlag.get())
-      throw new IterationInterruptedException();
-
     if (iter.hasNext()) {
       entry = iter.next();
       if (range.afterEndKey(entry.getKey())) {
@@ -110,9 +92,6 @@ public class SortedMapIterator implements InterruptibleIterator {
   @Override
   public void seek(Range range, Collection<ByteSequence> columnFamilies, boolean inclusive)
       throws IOException {
-
-    if (interruptFlag != null && interruptFlag.get())
-      throw new IterationInterruptedException();
 
     this.range = range;
 
@@ -139,10 +118,5 @@ public class SortedMapIterator implements InterruptibleIterator {
   public void init(SortedKeyValueIterator<Key,Value> source, Map<String,String> options,
       IteratorEnvironment env) throws IOException {
     throw new UnsupportedOperationException();
-  }
-
-  @Override
-  public void setInterruptFlag(AtomicBoolean flag) {
-    this.interruptFlag = flag;
   }
 }
