+++ server/tserver/src/main/java/org/apache/accumulo/tserver/InMemoryMap.java
@@ -53,10 +53,10 @@ import org.apache.accumulo.core.file.rfile.RFile;
 import org.apache.accumulo.core.file.rfile.RFileOperations;
 import org.apache.accumulo.core.iterators.IteratorEnvironment;
 import org.apache.accumulo.core.iterators.SortedKeyValueIterator;
-import org.apache.accumulo.core.iterators.SortedMapIterator;
 import org.apache.accumulo.core.iterators.WrappingIterator;
 import org.apache.accumulo.core.iteratorsImpl.system.EmptyIterator;
 import org.apache.accumulo.core.iteratorsImpl.system.InterruptibleIterator;
+import org.apache.accumulo.core.iteratorsImpl.system.InterruptibleMapIterator;
 import org.apache.accumulo.core.iteratorsImpl.system.LocalityGroupIterator;
 import org.apache.accumulo.core.iteratorsImpl.system.LocalityGroupIterator.LocalityGroup;
 import org.apache.accumulo.core.iteratorsImpl.system.SourceSwitchingIterator;
@@ -397,7 +397,7 @@ public class InMemoryMap {
       if (map == null)
         throw new IllegalStateException();
 
-      return new SortedMapIterator(map);
+      return new InterruptibleMapIterator(map, null);
     }
 
     @Override
